+++ examples/simple-examples/src/conversation/events/sendReadMessageEvent.ts
@@ -1,0 +1,29 @@
+import { Conversation } from '@sinch/sdk-core';
+import { getAppIdFromConfig, getContactIdFromConfig, initConversationService, printFullResponse } from '../../config';
+
+(async () => {
+  console.log('********************');
+  console.log('* Events_SendEvent *');
+  console.log('********************');
+
+  const appId = getAppIdFromConfig();
+  const contactId = getContactIdFromConfig();
+
+  const requestData: Conversation.SendReadMessageEventRequestData<Conversation.ContactId> = {
+    sendEventRequestBody: {
+      app_id: appId,
+      recipient: {
+        contact_id: contactId,
+      },
+      event: {
+        read_message_event: {},
+      },
+    },
+  };
+
+  const conversationService = initConversationService();
+  const response = await conversationService.events.sendReadMessageEvent(requestData);
+
+  printFullResponse(response);
+
+})();